+++ test/AuthorTest.php
@@ -4,7 +4,6 @@ namespace yii\akismet;
 
 use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
-use Psr\Http\Message\{UriInterface};
 
 /**
  * Tests the features of the `yii\akismet\Author` class.
@@ -22,7 +21,7 @@ class AuthorTest extends TestCase {
     it('should return an empty instance with an empty map', function() {
       $author = Author::fromJson([]);
       expect($author->email)->to->be->empty;
-      expect($author->url)->to->be->null;
+      expect($author->ipAddress)->to->be->empty;
     });
 
     it('should return an initialized instance with a non-empty map', function() {
@@ -40,37 +39,26 @@ class AuthorTest extends TestCase {
    * @test Author::jsonSerialize
    */
   public function testJsonSerialize() {
-    it('should return an empty map with a newly created instance', function() {
-      expect((new Author)->jsonSerialize())->to->be->empty;
+    it('should return only the IP address and user agent with a newly created instance', function() {
+      $data = (new Author('127.0.0.1', 'Doom/6.6.6'))->jsonSerialize();
+      expect(\Yii::getObjectVars($data))->to->have->lengthOf(2);
+      expect($data->user_agent)->to->equal('Doom/6.6.6');
+      expect($data->user_ip)->to->equal('127.0.0.1');
     });
 
     it('should return a non-empty map with a initialized instance', function() {
-      $data = (new Author([
+      $data = (new Author('192.168.0.1', 'Mozilla/5.0', [
         'email' => 'cedric@belin.io',
-        'ipAddress' => '127.0.0.1',
         'name' => 'Cédric Belin',
         'url' => 'https://belin.io'
       ]))->jsonSerialize();
 
+      expect(\Yii::getObjectVars($data))->to->have->lengthOf(5);
       expect($data->comment_author)->to->equal('Cédric Belin');
       expect($data->comment_author_email)->to->equal('cedric@belin.io');
       expect($data->comment_author_url)->to->equal('https://belin.io');
-      expect($data->user_ip)->to->equal('127.0.0.1');
-    });
-  }
-
-  /**
-   * @test Author::setUrl
-   */
-  public function testSetUrl() {
-    it('should return an instance of `UriInterface` for strings', function() {
-      $url = (new Author(['url' => 'https://github.com/cedx/yii2-akismet']))->url;
-      expect($url)->to->be->instanceOf(UriInterface::class);
-      expect((string) $url)->to->equal('https://github.com/cedx/yii2-akismet');
-    });
-
-    it('should return a `null` reference for unsupported values', function() {
-      expect((new Author(['url' => 123]))->url)->to->be->null;
+      expect($data->user_agent)->to->equal('Mozilla/5.0');
+      expect($data->user_ip)->to->equal('192.168.0.1');
     });
   }
 
@@ -78,9 +66,8 @@ class AuthorTest extends TestCase {
    * @test Author::__toString
    */
   public function testToString() {
-    $author = (string) new Author([
+    $author = (string) new Author('127.0.0.1', 'Doom/6.6.6', [
       'email' => 'cedric@belin.io',
-      'ipAddress' => '127.0.0.1',
       'name' => 'Cédric Belin',
       'url' => 'https://belin.io'
     ]);
@@ -93,6 +80,7 @@ class AuthorTest extends TestCase {
       expect($author)->to->contain('"comment_author":"Cédric Belin"')
         ->and->contain('"comment_author_email":"cedric@belin.io"')
         ->and->contain('"comment_author_url":"https://belin.io"')
+        ->and->contain('"user_agent":"Doom/6.6.6"')
         ->and->contain('"user_ip":"127.0.0.1"');
     });
   }