+++ test/CommentTest.php
@@ -4,9 +4,6 @@ namespace yii\akismet;
 
 use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
-use Psr\Http\Message\{UriInterface};
-use yii\base\{InvalidConfigException};
-use yii\validators\{Validator};
 
 /**
  * Tests the features of the `yii\akismet\Comment` class.
@@ -39,13 +36,11 @@ class CommentTest extends TestCase {
         'referrer' => 'https://belin.io'
       ]);
 
-      $author = $comment->author;
-      expect($author)->to->be->instanceOf(Author::class);
-      expect($author->name)->to->equal('Cédric Belin');
+      expect($comment->author)->to->be->instanceOf(Author::class);
+      expect($comment->author->name)->to->equal('Cédric Belin');
 
-      $date = $comment->date;
-      expect($date)->to->be->instanceOf(\DateTime::class);
-      expect($date->format('Y'))->to->equal(2000);
+      expect($comment->date)->to->be->instanceOf(\DateTime::class);
+      expect($comment->date->format('Y'))->to->equal(2000);
 
       expect($comment->content)->to->equal('A user comment.');
       expect($comment->referrer)->to->equal('https://belin.io');
@@ -54,109 +49,32 @@ class CommentTest extends TestCase {
   }
 
   /**
-   * @test Comment::isInstanceOf
-   */
-  public function testIsInstanceOf() {
-    it('should throw an exception if the `className` parameter is missing', function() {
-      expect(function() { (new Comment)->isInstanceOf('author', [], new Validator); })->to->throw(InvalidConfigException::class);
-    });
-
-    it('should throw an exception if the `className` parameter is an unknown class or interface', function() {
-      expect(function() { (new Comment)->isInstanceOf('author', ['className' => ''], new Validator); })->to->throw(InvalidConfigException::class);
-      expect(function() { (new Comment)->isInstanceOf('author', ['className' => 'Foo\Bar'], new Validator); })->to->throw(InvalidConfigException::class);
-    });
-
-    it('should add an error to the validator if the checked attribute is not an instance of the specified class or interface', function() {
-      $comment = new Comment(['author' => \Yii::createObject(Blog::class)]);
-      $comment->isInstanceOf('author', ['className' => Author::class], new Validator);
-      expect($comment->hasErrors())->to->be->true;
-    });
-
-    it('should not add any error to the validator if the checked attribute is an instance of the specified class or interface', function() {
-      $comment = new Comment(['author' => \Yii::createObject(Author::class)]);
-      $comment->isInstanceOf('author', ['className' => Author::class], new Validator);
-      expect($comment->hasErrors())->to->be->false;
-    });
-  }
-
-  /**
    * @test Comment::jsonSerialize
    */
   public function testJsonSerialize() {
-    it('should return an empty map with a newly created instance', function() {
-      expect((new Comment)->jsonSerialize())->to->be->empty;
+    it('should return only the author info with a newly created instance', function() {
+      $data = (new Comment(new Author('127.0.0.1', 'Doom/6.6.6')))->jsonSerialize();
+      expect(\Yii::getObjectVars($data))->to->have->lengthOf(2);
+      expect($data->user_agent)->to->equal('Doom/6.6.6');
+      expect($data->user_ip)->to->equal('127.0.0.1');
     });
 
     it('should return a non-empty map with a initialized instance', function() {
-      $data = (new Comment([
-        'author' => \Yii::createObject(['class' => Author::class, 'name' => 'Cédric Belin']),
+      $data = (new Comment(new Author('127.0.0.1', 'Doom/6.6.6', ['name' => 'Cédric Belin']), [
         'content' => 'A user comment.',
+        'date' => '2000-01-01T00:00:00.000Z',
         'referrer' => 'https://belin.io',
         'type' => CommentType::PINGBACK
       ]))->jsonSerialize();
 
+      expect(\Yii::getObjectVars($data))->to->have->lengthOf(7);
       expect($data->comment_author)->to->equal('Cédric Belin');
       expect($data->comment_content)->to->equal('A user comment.');
-      expect($data->comment_type)->to->equal(CommentType::PINGBACK);
+      expect($data->comment_date_gmt)->to->equal('2000-01-01T00:00:00+00:00');
+      expect($data->comment_type)->to->equal('pingback');
       expect($data->referrer)->to->equal('https://belin.io');
-    });
-  }
-
-  /**
-   * @test Comment::setDate
-   */
-  public function testSetDate() {
-    it('should return an instance of `DateTime` for strings and timestamps', function() {
-      expect((new Comment(['date' => time()]))->date)->to->be->instanceOf(\DateTime::class);
-      expect((new Comment(['date' => '2000-01-01T00:00:00+00:00']))->date)->to->be->instanceOf(\DateTime::class);
-    });
-
-    it('should return a `null` reference for unsupported values', function() {
-      expect((new Comment(['date' => []]))->date)->to->be->null;
-    });
-  }
-
-  /**
-   * @test Comment::setPermalink
-   */
-  public function testSetPermalink() {
-    it('should return an instance of `UriInterface` for strings', function() {
-      $url = (new Comment(['permalink' => 'https://github.com/cedx/yii2-akismet']))->permalink;
-      expect($url)->to->be->instanceOf(UriInterface::class);
-      expect((string) $url)->to->equal('https://github.com/cedx/yii2-akismet');
-    });
-
-    it('should return a `null` reference for unsupported values', function() {
-      expect((new Comment(['permalink' => 123]))->permalink)->to->be->null;
-    });
-  }
-
-  /**
-   * @test Comment::setPostModified
-   */
-  public function testSetPostModified() {
-    it('should return an instance of `DateTime` for strings and timestamps', function() {
-      expect((new Comment(['postModified' => time()]))->postModified)->to->be->instanceOf(\DateTime::class);
-      expect((new Comment(['postModified' => '2000-01-01T00:00:00+00:00']))->postModified)->to->be->instanceOf(\DateTime::class);
-    });
-
-    it('should return a `null` reference for unsupported values', function() {
-      expect((new Comment(['postModified' => []]))->postModified)->to->be->null;
-    });
-  }
-
-  /**
-   * @test Comment::setReferrer
-   */
-  public function testSetReferrer() {
-    it('should return an instance of `UriInterface` for strings', function() {
-      $url = (new Comment(['referrer' => 'https://github.com/cedx/yii2-akismet']))->referrer;
-      expect($url)->to->be->instanceOf(UriInterface::class);
-      expect((string) $url)->to->equal('https://github.com/cedx/yii2-akismet');
-    });
-
-    it('should return a `null` reference for unsupported values', function() {
-      expect((new Comment(['referrer' => 123]))->referrer)->to->be->null;
+      expect($data->user_agent)->to->equal('Doom/6.6.6');
+      expect($data->user_ip)->to->equal('127.0.0.1');
     });
   }
 
@@ -164,9 +82,9 @@ class CommentTest extends TestCase {
    * @test Comment::__toString
    */
   public function testToString() {
-    $comment = (string) new Comment([
-      'author' => \Yii::createObject(['class' => Author::class, 'name' => 'Cédric Belin']),
+    $comment = (string) new Comment(new Author('127.0.0.1', 'Doom/6.6.6', ['name' => 'Cédric Belin']), [
       'content' => 'A user comment.',
+      'date' => '2000-01-01T00:00:00.000Z',
       'referrer' => 'https://belin.io',
       'type' => CommentType::PINGBACK
     ]);
@@ -179,7 +97,10 @@ class CommentTest extends TestCase {
       expect($comment)->to->contain('"comment_author":"Cédric Belin"')
         ->and->contain('"comment_content":"A user comment."')
         ->and->contain('"comment_type":"pingback"')
-        ->and->contain('"referrer":"https://belin.io"');
+        ->and->contain('"comment_date_gmt":"2000-01-01T00:00:00+00:00"')
+        ->and->contain('"referrer":"https://belin.io"')
+        ->and->contain('"user_agent":"Doom/6.6.6"')
+        ->and->contain('"user_ip":"127.0.0.1"');
     });
   }
 }